Create common dataclasses and refactor agents
